add description to user menu

+++ app/controllers/api/user_menus_controller.rb
@@ -9,13 +9,21 @@ class Api::UserMenusController < Api::ApplicationController
   def update
     user_menu = UserMenu.find(params[:id])
 
-    user_menu.dishes.delete_all
-    user_menu.dishes << Dish.where(id: dishes_params.map { |dish| dish[:id] })
+    if user_menu.update(user_menu_params)
+      user_menu.dishes.delete_all
+      user_menu.dishes << Dish.where(id: dishes_params.map { |dish| dish[:id] })
 
-    render json: user_menu
+      render json: user_menu
+    else
+      render json: { errors: user_menu.errors }
+    end
   end
 
   private
+
+  def user_menu_params
+    params.require(:user_menu).permit(:description)
+  end
 
   def dishes_params
     params.require(:user_menu).permit(dishes: [:id])[:dishes] || []
